requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.13  # optional, only used if OCR toggle is enabled and tesseract is installed
qrcode==7.4.2
